chat_companion/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chat-companion"
version = "0.1.0"
description = "Companion chatbot toolkit with OpenRouter LLM client, prompt templates, and layered memory modules."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "Chat Companion Maintainers", email = "maintainers@example.com" }
]
dependencies = [
  "httpx>=0.27",
  "pydantic>=2.5",
  "python-dotenv>=1.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-asyncio>=0.23"
]

[project.scripts]
chat-companion = "chat_companion.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
